refactor: less api duplication
