+++ enum-collections/src/enummap.rs
@@ -1,5 +1,5 @@
-mod index;
 use std::marker::PhantomData;
+use std::ops::{Index, IndexMut};
 
 use crate::Enumerated;
 
@@ -100,16 +100,46 @@ where
     }
 }
 
+impl<K, V> Index<K> for EnumMap<K, V>
+where
+    K: Enumerated,
+    V: Default,
+{
+    type Output = Option<V>;
+
+    fn index(&self, key: K) -> &Self::Output {
+        &self.values[key.position()]
+    }
+}
+
+impl<K, V> IndexMut<K> for EnumMap<K, V>
+where
+    K: Enumerated,
+    V: Default,
+{
+    fn index_mut(&mut self, key: K) -> &mut Self::Output {
+        &mut self.values[key.position()]
+    }
+}
+
 #[cfg(test)]
 mod tests {
-    use crate::Enumerated;
-
     use super::EnumMap;
+    use crate::Enumerated;
 
     #[derive(Enumerated)]
     pub(super) enum Letter {
         A,
         B,
+    }
+
+    #[test]
+    fn get_insert_index_trait() {
+        let mut enum_map = EnumMap::<Letter, i32>::new();
+        enum_map[Letter::A] = Some(42);
+        assert_eq!(Some(42), enum_map[Letter::A]);
+        assert_eq!(Some(&42), enum_map[Letter::A].as_ref());
+        assert_eq!(None, enum_map[Letter::B]);
     }
 
     #[test]
